+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/mining/forging/Forging.java
@@ -1,5 +1,7 @@
 package insane96mcp.iguanatweaksexpanded.module.mining.forging;
 
+import insane96mcp.iguanatweaksexpanded.InsaneSurvivalExtra;
+import insane96mcp.iguanatweaksexpanded.data.generator.ISEItemTagsProvider;
 import insane96mcp.iguanatweaksexpanded.module.Modules;
 import insane96mcp.iguanatweaksexpanded.module.items.copper.CopperExpansion;
 import insane96mcp.iguanatweaksexpanded.module.items.solarium.Solarium;
@@ -11,22 +13,36 @@ import insane96mcp.iguanatweaksexpanded.module.misc.ISEDataPacks;
 import insane96mcp.iguanatweaksexpanded.setup.ISERegistries;
 import insane96mcp.iguanatweaksexpanded.setup.IntegratedPack;
 import insane96mcp.iguanatweaksexpanded.setup.registry.SimpleBlockWithItem;
+import insane96mcp.iguanatweaksreborn.event.ISOLivingAttackEvent;
+import insane96mcp.iguanatweaksreborn.module.experience.enchantments.EnchantmentsFeature;
 import insane96mcp.iguanatweaksreborn.module.items.flintexpansion.FlintExpansion;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.LoadFeature;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
+import net.minecraft.core.particles.BlockParticleOption;
+import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.network.chat.Component;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.packs.PackType;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.ai.attributes.Attributes;
+import net.minecraft.world.entity.decoration.ArmorStand;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.flag.FeatureFlags;
 import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Tiers;
 import net.minecraft.world.item.crafting.RecipeType;
+import net.minecraft.world.item.enchantment.EnchantmentHelper;
+import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockBehaviour;
+import net.minecraftforge.event.entity.living.LivingKnockBackEvent;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.registries.RegistryObject;
 
 @Label(name = "Forging")
@@ -76,5 +92,45 @@ public class Forging extends Feature {
 	public Forging(Module module, boolean enabledByDefault, boolean canBeDisabled) {
 		super(module, enabledByDefault, canBeDisabled);
 		IntegratedPack.addPack(new IntegratedPack(PackType.SERVER_DATA, "forging_equipment", Component.literal("IguanaTweaks Expanded Forging Equipment"), () -> this.isEnabled() && !ISEDataPacks.disableAllDataPacks && forgingEquipment));
+	}
+
+	@SubscribeEvent
+	public void onHammerDamage(ISOLivingAttackEvent event) {
+		if (!(event.getSource().getEntity() instanceof LivingEntity attacker)
+				|| !attacker.getMainHandItem().is(ISEItemTagsProvider.FORGE_HAMMERS)
+				|| event.getEntity().level().isClientSide)
+			return;
+
+		event.getEntity().getPersistentData().putBoolean(InsaneSurvivalExtra.MOD_ID + "cancel_knockback", true);
+
+		if ((attacker instanceof Player player && player.getAttackStrengthScale(0.5f) < 0.9f))
+			return;
+
+		float range = 2.5F;
+		float rangeSqr = range * range;
+
+		for (LivingEntity livingEntity : event.getEntity().level().getEntitiesOfClass(LivingEntity.class, event.getEntity().getBoundingBox().inflate(range, range / 2f, range))) {
+			if (livingEntity != attacker
+					&& !livingEntity.isAlliedTo(attacker)
+					&& (!(livingEntity instanceof ArmorStand armorStand) || !armorStand.isMarker())
+					&& event.getEntity().distanceToSqr(livingEntity) < rangeSqr) {
+				livingEntity.push(0, (1f + (getKnockbackBonus(attacker))) * (1.0D - livingEntity.getAttributeValue(Attributes.KNOCKBACK_RESISTANCE)), 0);
+			}
+		}
+
+		event.getEntity().playSound(SoundEvents.ANVIL_PLACE, 0.6f, 1.1f);
+		((ServerLevel) event.getEntity().level()).sendParticles(new BlockParticleOption(ParticleTypes.BLOCK, Blocks.ANVIL.defaultBlockState), event.getEntity().getX(), event.getEntity().getY(), event.getEntity().getZ(), 200, range / 2f, range / 4f, range / 2f, 1f);
+	}
+
+	private static float getKnockbackBonus(LivingEntity entity) {
+		return Math.max(EnchantmentHelper.getEnchantmentLevel(Enchantments.KNOCKBACK, entity), EnchantmentHelper.getEnchantmentLevel(EnchantmentsFeature.KNOCKBACK.get(), entity)) * 0.35f;
+	}
+
+	@SubscribeEvent
+	public void cancelKnockback(LivingKnockBackEvent event) {
+		if (event.getEntity().getPersistentData().getBoolean(InsaneSurvivalExtra.MOD_ID + "cancel_knockback")) {
+			event.setCanceled(true);
+			event.getEntity().getPersistentData().remove(InsaneSurvivalExtra.MOD_ID + "cancel_knockback");
+		}
 	}
 }